requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
beautifulsoup4
certifi
charset-normalizer
frozenlist
greenlet
idna
multidict
nest-asyncio
playwright
propcache
psycopg
psycopg-binary
pyee
python-dotenv
requests
soupsieve
typing_extensions
Unidecode
urllib3
yarl
